26_python-flask/flask_exercises-database/pyproject.toml
[project]
name = "hexlet-flask"
version = "0.1.0"
description = "Example application"
authors = ["paalso <40a.pvl@gmail.com>"]
readme = "README.md"
requires-python = ">=3.11.4"

dependencies = [
    "flask>=3.1.0",
    "faker>=33.1.0",
    "psycopg[binary]>=3.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.1",
]
